Agents/requirements.txt
langchain
wikipedia
python-dotenv
langchain-core
langchain-community
pydantic
langchain-openai
